requirements.txt
# Code Review Assistant - Python Dependencies

# Core Framework
streamlit>=1.28.0

# HTTP Requests
requests>=2.31.0

# PDF Generation
reportlab>=4.0.0

# Markdown Processing
markdown>=3.5.0

# Environment Variables
python-dotenv>=1.0.0

# Database
# sqlite3 is built-in with Python

# File Processing
python-magic>=0.4.27

# Additional Utilities
# uuid, datetime, json, os, pathlib, typing, traceback, re, time are built-in with Python

# Development Dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
